# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Yercaud Business Directory API
  version: 0.1.0
  description: >
    Backend contract for the public frontend and Admin panel. Auth is
    httpOnly session cookies (ADR: session cookies over JWT). See
    /CONTEXT.md and /docs/adr/ for the domain model and design decisions
    this contract implements.
servers:
  - url: http://localhost:4000/v1
    description: Local dev
security:
  - cookieAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Create a User account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string }
      responses:
        '201':
          description: Account created, session cookie set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Session cookie set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me:
    patch:
      tags: [Auth]
      summary: Update the current User's own profile
      description: >-
        Added in Phase 10 — nothing let a User edit their own profile.
        PATCH /users/{userId} is the Super Admin's suspend/reactivate, and /me/*
        was read-only.

        Self-scoped by construction: it acts on the session's User, so there is no
        id in the path to tamper with. Email is deliberately not editable — it's
        the login identity and the unique key, so changing it is an
        account-recovery flow, not a profile edit. Password has its own reset
        flow. Roles and status are not self-editable, and are stripped from the
        body rather than honoured.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OwnProfileInput' }
      responses:
        '200':
          description: OK — the same shape as /auth/me, so a client can swap it straight into its cache.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/username-available:
    get:
      tags: [Auth]
      summary: Live username-availability check while the User is typing (issue #14)
      description: >-
        Advisory only, ahead of a full form submit — PATCH /me is still the
        source of truth and re-validates on save. Excludes the caller's own
        current username, so keeping it never reads as taken.
      parameters:
        - name: username
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [available]
                properties:
                  available: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/avatar:
    post:
      tags: [Auth]
      summary: Upload a real profile photo (issue #14), replacing the client's blob:-URL workaround
      description: >-
        Reuses the on-disk storage and static-serving approach already used by
        listing images. JPEG/PNG/WEBP only, 2 MB cap.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: Uploaded — same shape as /auth/me, with the new avatarUrl already applied.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/stats:
    get:
      tags: [Auth]
      summary: The current User's own activity counts, for the profile page's My Statistics card (issue #13)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [enquiries, reviews, favorites, listings]
                properties:
                  enquiries: { type: integer }
                  reviews: { type: integer }
                  favorites: { type: integer }
                  listings:
                    type: integer
                    description: Listings across all Businesses this User owns — 0 for a non-owner, not omitted.
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/password:
    post:
      tags: [Auth]
      summary: Change (or, for a Google-only account, set a first) password while signed in — issue #13
      description: >-
        Distinct from the forgot-password email/token flow, which is
        unchanged. currentPassword is required and verified when the account
        already has one; for a Google-only account (hasPassword: false on
        CurrentUser) it's ignored, since there's nothing to verify.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newPassword]
              additionalProperties: false
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        '204': { description: Password changed }
        '401': { description: Not signed in, or currentPassword missing/incorrect when one is required }
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Current session's User, with roles and effective permissions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # Phase 6: not in the original Phase 0.5 contract — POST /users/{userId}/impersonate
  # starts a session but nothing stops it. Self-scoped: exits whatever
  # impersonation session the caller is currently on.
  /auth/exit-impersonation:
    post:
      tags: [Auth]
      summary: Exit an active impersonation session, restoring the admin's own session (FR160)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /auth/password-reset/request:
    post:
      tags: [Auth]
      security: []
      summary: Send a password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties: { email: { type: string, format: email } }
      responses:
        '202': { description: Reset email queued if the account exists }
  /auth/password-reset/confirm:
    post:
      tags: [Auth]
      security: []
      summary: Complete a password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        '204': { description: Password updated }
        '400': { $ref: '#/components/responses/BadRequest' }

  # ===========================================================================
  # Taxonomies (Super Admin write, public read)
  # ===========================================================================
...
  /users/{userId}/impersonate:
    parameters:
      - name: userId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Admin]
      summary: Super Admin "view as" a Business Owner (FR160) — recorded in audit log
      responses:
        '200':
          description: Impersonation session started
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ===========================================================================
  # Marketing
  # ===========================================================================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Not permitted (RBAC)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflict:
      description: Conflicts with an existing resource or constraint
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
...
    User:
      type: object
      required: [id, email, name, username, phone, avatarUrl, bio, dateOfBirth, gender, status, roles, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        username:
          type: string
          pattern: '^[a-z][a-z0-9_]{2,19}$'
          description: >-
            Public handle (issue #14), auto-generated at registration and
            renameable via PATCH /me. Shown alongside name wherever a Review or
            Enquiry author is displayed to someone other than themselves.
        phone: { type: string, nullable: true }
        avatarUrl: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        dateOfBirth:
          type: string
          format: date
          nullable: true
          description: Storage-only (issue #14) — nothing in the app reads or acts on this.
        gender:
          type: string
          nullable: true
          enum: [Female, Male, Non-binary, 'Prefer not to say', null]
          description: Storage-only (issue #14).
        status: { type: string, enum: [active, suspended, deleted] }
        roles: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }
    # Phase 6: used only by GET /auth/me, POST /users/{userId}/impersonate,
    # and POST /auth/exit-impersonation — the plain User schema (GET /users,
    # etc.) stays unchanged so this doesn't leak an impersonation-specific
    # field into unrelated responses.
    OwnProfileInput:
      type: object
      additionalProperties: false
      description: Every field is optional; null clears a nullable field, absent leaves it untouched. username has no null clear — it's NOT NULL — but is otherwise optional/absent-leaves-alone.
      properties:
        name: { type: string, maxLength: 255 }
        username: { type: string, minLength: 3, maxLength: 20, pattern: '^[a-z][a-z0-9_]{2,19}$' }
        phone: { type: string, maxLength: 30, nullable: true }
        avatarUrl: { type: string, nullable: true }
        bio: { type: string, maxLength: 1000, nullable: true }
        dateOfBirth: { type: string, format: date, nullable: true }
        gender: { type: string, nullable: true, enum: [Female, Male, Non-binary, 'Prefer not to say', null] }
    CurrentUser:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          required: [permissions, impersonatedBy, emailVerifiedAt, hasPassword]
          properties:
            impersonatedBy: { type: string, format: uuid, nullable: true }
            emailVerifiedAt: { type: string, format: date-time, nullable: true }
            hasPassword:
              type: boolean
              description: >-
                False for a Google-only account (issue #13) that has never set a
                password. Drives whether Account Settings shows "Change
                Password" or "Set Password".
            permissions:
              type: array
              items: { $ref: '#/components/schemas/Permission' }
              description: >-
                The caller's effective permissions, unioned across their Roles and
                deduplicated (Phase 10). This is what lets the Admin panel render
                navigation and actions for a Super-Admin-created custom Role with no
                frontend change — branching on Role names would leave a "Content
                Moderator" seeing the Business Owner UI or nothing.

                Advisory, for rendering only. Every endpoint keeps its own guard;
                hiding a button is a courtesy, not a control. Empty for a User with
                no Roles, which is a real account (a plain public visitor), not an
                error.
...
    Permission:
      type: object
      required: [module, action]
      properties:
        module: { type: string }
        action: { type: string, enum: [view, create, edit, delete, approve, publish] }
...